add example

+++ model/layer/utils.py
@@ -61,6 +61,6 @@ def get_extended_attention_mask(attention_mask: Tensor, input_shape: Tuple[int],
       # positions we want to attend and -10000.0 for masked positions.
       # Since we are adding it to the raw scores before the softmax, this is
       # effectively the same as removing these entirely.
-      extended_attention_mask = extended_attention_mask.to(dtype=torch.float16)  # fp16 compatibility
+    
       extended_attention_mask = (1.0 - extended_attention_mask) * -10000.0
       return extended_attention_mask
